requirements.txt
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
playwright==1.50.0
playwright-stealth==1.0.6
pyee==12.1.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
setuptools==75.8.2
singlestoredb==1.12.1
six==1.17.0
sqlparams==6.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
